server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymongo==4.6.0
PyPDF2==3.0.1
pandas>=2.1.3
scikit-learn>=1.3.2
nltk==3.8.1
python-dotenv==1.0.0
krutrim-cloud>=0.7.1